refactor: extract useColumnOrder hook from ItemsTable
Pulls the localStorage init, participant sync, and persist effects out of
ItemsTable into a reusable useColumnOrder hook. ItemsTable now accepts
columnOrder and onReorderColumns as props; App calls the hook once and
passes the results down.

No behavior change — column drag reorder and localStorage persistence
work exactly as before.
